# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Vartalaap API",
    "description": "Voice bot platform for local Indian businesses",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/reservations": {
      "get": {
        "tags": [
          "CRUD",
          "Reservations"
        ],
        "summary": "List Reservations",
        "description": "List reservations with optional filters.\n\nSecurity: business_id is required to prevent cross-tenant data access.",
        "operationId": "list_reservations_api_reservations_get",
        "parameters": [
          {
            "name": "business_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Required for tenant isolation",
              "title": "Business Id"
            },
            "description": "Required for tenant isolation"
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/ReservationStatus"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "name": "date_from",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Date From"
            }
          },
          {
            "name": "date_to",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Date To"
            }
          },
          {
            "name": "skip",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Skip"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 1000,
              "minimum": 1,
              "default": 100,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ReservationResponse"
                  },
                  "title": "Response List Reservations Api Reservations Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "CRUD",
          "Reservations"
        ],
        "summary": "Create Reservation",
        "description": "Create a new reservation.",
        "operationId": "create_reservation_api_reservations_post",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReservationCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReservationResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/reservations/{reservation_id}": {
      "get": {
        "tags": [
          "CRUD",
          "Reservations"
        ],
        "summary": "Get Reservation",
        "description": "Get a reservation by ID.",
        "operationId": "get_reservation_api_reservations__reservation_id__get",
        "parameters": [
          {
            "name": "reservation_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Reservation Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReservationResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "CRUD",
          "Reservations"
        ],
        "summary": "Update Reservation",
        "description": "Update a reservation (partial update).",
        "operationId": "update_reservation_api_reservations__reservation_id__patch",
        "parameters": [
          {
            "name": "reservation_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Reservation Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReservationUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReservationResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "CRUD",
          "Reservations"
        ],
        "summary": "Delete Reservation",
        "description": "Delete a reservation.",
        "operationId": "delete_reservation_api_reservations__reservation_id__delete",
        "parameters": [
          {
            "name": "reservation_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Reservation Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ReservationCreate": {
        "properties": {
          "business_id": {
            "type": "string",
            "title": "Business Id"
          },
          "call_log_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Call Log Id"
          },
          "customer_name": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 100
              },
              {
                "type": "null"
              }
            ],
            "title": "Customer Name"
          },
          "customer_phone_encrypted": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Customer Phone Encrypted"
          },
          "party_size": {
            "type": "integer",
            "maximum": 20.0,
            "minimum": 1.0,
            "title": "Party Size"
          },
          "reservation_date": {
            "type": "string",
            "format": "date",
            "title": "Reservation Date"
          },
          "reservation_time": {
            "type": "string",
            "pattern": "^([01]?[0-9]|2[0-3]):[0-5][0-9]$",
            "title": "Reservation Time"
          },
          "notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          },
          "whatsapp_consent": {
            "type": "boolean",
            "title": "Whatsapp Consent",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "business_id",
          "party_size",
          "reservation_date",
          "reservation_time"
        ],
        "title": "ReservationCreate",
        "description": "Schema for creating a reservation."
      },
      "ReservationResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "business_id": {
            "type": "string",
            "title": "Business Id"
          },
          "call_log_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Call Log Id"
          },
          "customer_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Customer Name"
          },
          "party_size": {
            "type": "integer",
            "title": "Party Size"
          },
          "reservation_date": {
            "type": "string",
            "title": "Reservation Date"
          },
          "reservation_time": {
            "type": "string",
            "title": "Reservation Time"
          },
          "status": {
            "$ref": "#/components/schemas/ReservationStatus"
          },
          "whatsapp_sent": {
            "type": "boolean",
            "title": "Whatsapp Sent"
          },
          "whatsapp_consent": {
            "type": "boolean",
            "title": "Whatsapp Consent"
          },
          "notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          },
          "created_at": {
            "type": "string",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "business_id",
          "call_log_id",
          "customer_name",
          "party_size",
          "reservation_date",
          "reservation_time",
          "status",
          "whatsapp_sent",
          "whatsapp_consent",
          "notes",
          "created_at",
          "updated_at"
        ],
        "title": "ReservationResponse",
        "description": "Response schema for reservations."
      },
...
      "ReservationStatus": {
        "type": "string",
        "enum": [
          "confirmed",
          "cancelled",
          "completed",
          "no_show"
        ],
        "title": "ReservationStatus",
        "description": "Status of a reservation."
      },
      "ReservationUpdate": {
        "properties": {
          "customer_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Customer Name"
          },
          "party_size": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 20.0,
                "minimum": 1.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Party Size"
          },
          "reservation_date": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reservation Date"
          },
          "reservation_time": {
            "anyOf": [
              {
                "type": "string",
                "pattern": "^([01]?[0-9]|2[0-3]):[0-5][0-9]$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reservation Time"
          },
          "status": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ReservationStatus"
              },
              {
                "type": "null"
              }
            ]
          },
          "notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          },
          "whatsapp_sent": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Whatsapp Sent"
          }
        },
        "type": "object",
        "title": "ReservationUpdate",
        "description": "Schema for updating a reservation (all fields optional)."
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}